+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/RedirectUriClientTest.kt
@@ -12,12 +12,14 @@ package at.asitplus.wallet.lib.openid
  * see the "LICENSE" file for more details
  */
 
+import at.asitplus.data.NonEmptyList.Companion.toNonEmptyList
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.testballoon.matrix.fixture
 import at.asitplus.testballoon.matrix.matrixSuite
 import at.asitplus.wallet.lib.RequestOptionsCredential
+import at.asitplus.openid.VerifierInfo
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.Holder
 import at.asitplus.wallet.lib.agent.HolderAgent
@@ -31,6 +33,8 @@ import at.asitplus.wallet.lib.oidvci.encodeToParameters
 import at.asitplus.wallet.lib.oidvci.formUrlEncode
 import at.asitplus.wallet.lib.openid.DummyCredentialDataProvider.issueAndStorePlainJwt
 import at.asitplus.wallet.lib.utils.MapStore
+import at.asitplus.signum.indispensable.pki.X509Certificate
+import at.asitplus.wallet.lib.NonceService
 import com.benasher44.uuid.uuid4
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
@@ -95,6 +99,74 @@ val RedirectUriClientTest by matrixSuite {
             verifySecondProtocolRun(it.verifierOid4vp, it.walletUrl, it.holderOid4vp)
         }
 
+        "verifier_info is exposed in preparation state" {
+            val verifierInfo = listOf(
+                VerifierInfo(
+                    format = "jwt",
+                    data = "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9",
+                    credentialIds = setOf("id_card"),
+                )
+            ).toNonEmptyList()
+
+            val requestOptions = OpenId4VpRequestOptions(
+                presentationRequest = CredentialPresentationRequestBuilder(
+                    credentials = setOf(
+                        RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
+                    )
+                ).toPresentationExchangeRequest(),
+                verifierInfo = verifierInfo
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(
+                requestOptions, CreationOptions.Query(it.walletUrl)
+            ).getOrThrow().url
+
+            val preparationState = it.holderOid4vp
+                .startAuthorizationResponsePreparation(authnRequest)
+                .getOrThrow()
+
+            preparationState.verifierInfo shouldBe verifierInfo
+        }
+
+        "registration_cert verifier_info is exposed in preparation state" {
+            val verifierInfo = listOf(
+                VerifierInfo(
+                    format = OpenIdConstants.VerifierInfo.REGISTRATION_CERT_FORMAT,
+                    data = "eyJhbGciOiJFUzI1NiJ9.eyJzdWIiOiJ3cnAifQ.signature",
+                    credentialIds = setOf("id_card"),
+                )
+            ).toNonEmptyList()
+            val requestOptions = OpenId4VpRequestOptions(
+                presentationRequest = CredentialPresentationRequestBuilder(
+                    credentials = setOf(
+                        RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
+                    )
+                ).toPresentationExchangeRequest(),
+                verifierInfo = verifierInfo,
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(
+                requestOptions, CreationOptions.Query(it.walletUrl)
+            ).getOrThrow().url
+
+            val preparationState = it.holderOid4vp
+                .startAuthorizationResponsePreparation(authnRequest)
+                .getOrThrow()
+
+            preparationState.verifierInfo shouldBe verifierInfo
+        }
+
+        "certificateSanDnsFromPem parses PEM chain and builds client_id" {
+            val clientIdScheme = ClientIdScheme.CertificateSanDns(
+                chain = parsePemCertificateChain(samplePemChain),
+                clientIdDnsName = "*.google.com",
+                redirectUri = "https://example.com/callback",
+            )
+
+            clientIdScheme.clientIdWithoutPrefix shouldBe "*.google.com"
+            clientIdScheme.clientId shouldBe "x509_san_dns:*.google.com"
+            clientIdScheme.redirectUri shouldBe "https://example.com/callback"
+            clientIdScheme.chain.size shouldBe 2
+        }
+        
         "wrong client nonce in vp_token should lead to error" {
             val verifierOid4vp = OpenId4VpVerifier(
                 keyMaterial = it.verifierKeyMaterial,
@@ -323,3 +395,77 @@ private val defaultRequestOptions = OpenId4VpRequestOptions(
         RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
     ).toDCQLRequest(),
 )
+
+private val samplePemChain = listOf(
+    """
+        -----BEGIN CERTIFICATE-----
+MIIGvTCCBaWgAwIBAgIIAb1LAwRonlswDQYJKoZIhvcNAQELBQAwcTELMAkGA1UE
+BhMCR0IxDzANBgNVBAgTBkxvbmRvbjEPMA0GA1UEBxMGTG9uZG9uMQ8wDQYDVQQK
+EwZHb29nbGUxDDAKBgNVBAsTA0VuZzEhMB8GA1UEAxMYRmFrZUNlcnRpZmljYXRl
+QXV0aG9yaXR5MB4XDTIzMDgxMTA2NDg0MVoXDTI2MDQxMjA2NDg0MVowggE2MQsw
+CQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwNTW91bnRh
+aW4gVmlldzETMBEGA1UECgwKR29vZ2xlIEluYzEVMBMGA1UEAwwMKi5nb29nbGUu
+Y29tMYGYMIGVBgNVBAQMgY1SRkM1MjgwIHM0LjIgJ2FwcGxpY2F0aW9ucyBjb25m
+b3JtaW5nIHRvIHRoaXMgcHJvZmlsZSBNVVNUIHJlY29nbml6ZSB0aGUgZm9sbG93
+aW5nIGV4dGVuc2lvbnM6IC4uLnN1YmplY3QgYWx0ZXJuYXRpdmUgbmFtZSAoU2Vj
+dGlvbiA0LjIuMS42KScxMzAxBgNVBCoMKkluY2x1ZGUgU3ViamVjdCBBbHRlcm5h
+dGl2ZSBOYW1lIGV4dGVuc2lvbjBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABLrd
+rvPwAwPVXyJV4b8dFs3wwB6EdvlwIjwasxr5SSTEoBQp9rGgLkuGO59x860UnvLR
+xKjYFEOKsOt4zjejJBCjggNbMIIDVzAPBgNVHSMECDAGgAQBAgMEMIIDQgYDVR0R
+BIIDOTCCAzWCDCouZ29vZ2xlLmNvbYINKi5hbmRyb2lkLmNvbYIWKi5hcHBlbmdp
+bmUuZ29vZ2xlLmNvbYISKi5jbG91ZC5nb29nbGUuY29tghYqLmdvb2dsZS1hbmFs
+eXRpY3MuY29tggsqLmdvb2dsZS5jYYILKi5nb29nbGUuY2yCDiouZ29vZ2xlLmNv
+Lmlugg4qLmdvb2dsZS5jby5qcIIOKi5nb29nbGUuY28udWuCDyouZ29vZ2xlLmNv
+bS5hcoIPKi5nb29nbGUuY29tLmF1gg8qLmdvb2dsZS5jb20uYnKCDyouZ29vZ2xl
+LmNvbS5jb4IPKi5nb29nbGUuY29tLm14gg8qLmdvb2dsZS5jb20udHKCDyouZ29v
+Z2xlLmNvbS52boILKi5nb29nbGUuZGWCCyouZ29vZ2xlLmVzggsqLmdvb2dsZS5m
+coILKi5nb29nbGUuaHWCCyouZ29vZ2xlLml0ggsqLmdvb2dsZS5ubIILKi5nb29n
+bGUucGyCCyouZ29vZ2xlLnB0ghIqLmdvb2dsZWFkYXBpcy5jb22CDyouZ29vZ2xl
+YXBpcy5jboIUKi5nb29nbGVjb21tZXJjZS5jb22CESouZ29vZ2xldmlkZW8uY29t
+ggwqLmdzdGF0aWMuY26CDSouZ3N0YXRpYy5jb22CCiouZ3Z0MS5jb22CCiouZ3Z0
+Mi5jb22CFCoubWV0cmljLmdzdGF0aWMuY29tggwqLnVyY2hpbi5jb22CECoudXJs
+Lmdvb2dsZS5jb22CFioueW91dHViZS1ub2Nvb2tpZS5jb22CDSoueW91dHViZS5j
+b22CFioueW91dHViZWVkdWNhdGlvbi5jb22CCyoueXRpbWcuY29tghphbmRyb2lk
+LmNsaWVudHMuZ29vZ2xlLmNvbYILYW5kcm9pZC5jb22CBGcuY2+CBmdvby5nbIIU
+Z29vZ2xlLWFuYWx5dGljcy5jb22CCmdvb2dsZS5jb22CEmdvb2dsZWNvbW1lcmNl
+LmNvbYIKdXJjaGluLmNvbYIIeW91dHUuYmWCC3lvdXR1YmUuY29tghR5b3V0dWJl
+ZWR1Y2F0aW9uLmNvbTANBgkqhkiG9w0BAQsFAAOCAQEAg0dBdGdXcsfQcXKGB7F6
+iOdKJpbVMYwtH43hWNdLbdReMQ5h7eV1y5Vhra5ZJkkWX71v12AEZkD3DpunQ7g7
+FP2xPDY2r/Hmw61Zzqm2aEvrAg4PEkypqkfQ6/NiVspqBF3zGinm9qYp/Ifl+vf1
+Et9TbTIluVQa2DylQKQ0cKKvQnGsleuCWuxpbJmEZu7q/nMNUwbrs5Ln/FBKMLkJ
++jvuiS71FFeiAFhvXa48j2BoWh9Rpots8iiF4iuGbol8z9NAojwarKyLuTXuGAhv
+IBH+LmnJQP5yoUjjgzleBHCD5ENQ4md2eii35cszAkwWOdVGXQpvRnwL2NSfdBe+
+cQ==
+-----END CERTIFICATE-----
+    """,
+    """
+        -----BEGIN CERTIFICATE-----
+MIIDpzCCAo+gAwIBAgIEBAbK/jANBgkqhkiG9w0BAQsFADBxMQswCQYDVQQGEwJH
+QjEPMA0GA1UECBMGTG9uZG9uMQ8wDQYDVQQHEwZMb25kb24xDzANBgNVBAoTBkdv
+b2dsZTEMMAoGA1UECxMDRW5nMSEwHwYDVQQDExhGYWtlQ2VydGlmaWNhdGVBdXRo
+b3JpdHkwHhcNMjMxMDEwMDY0ODQwWhcNMjQxMDA5MDY0ODQwWjBxMQswCQYDVQQG
+EwJHQjEPMA0GA1UECBMGTG9uZG9uMQ8wDQYDVQQHEwZMb25kb24xDzANBgNVBAoT
+Bkdvb2dsZTEMMAoGA1UECxMDRW5nMSEwHwYDVQQDExhGYWtlQ2VydGlmaWNhdGVB
+dXRob3JpdHkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDWH5ATQjwc
+3aEmsFpltt2PE/hb/Ff9RIaD5W6+Kbi29Q//H4YZFGbk8BVokVuD5yxPsF9aCGfz
+T1YqdBtpl1r5Sn9A6BlfSguS6rVDN/zq3L7/OHFIBKF28nibyrxd3sd5wBH2X+AB
+c6FU2rP2EeV2T+tCvWNukOH/M6fxkhITKrcJ9ZdFT36MCAsjklBxV/pXzAX3ZXBp
+i/2pvYmHqgIlnadCONylUmGn5XY0h9gY/6PUASMBiPyhj1wcwguWgdBm0nwhJ51m
+VezRTUs45xs9cbpKMiqkSuHR/kiPqnAfDawyf6KsZhiSCNIZfYZT7Mh4m4a5F2sn
+/pZHFn6Kz6SbAgMBAAGjRzBFMA0GA1UdDgQGBAQBAgMEMA8GA1UdIwQIMAaABAEC
+AwQwEgYDVR0TAQH/BAgwBgEB/wIBCjAPBgNVHQ8BAf8EBQMDB/+AMA0GCSqGSIb3
+DQEBCwUAA4IBAQDKbgFp4SuOghVwFRAWI2f/gOBLIF1UQYzUGqfeVLZGS06xqqYO
+lX+RH5RmeY+eG/AtuCi4vgn6f9h0W/0L68sjbs4qIK9t/fBDaaQ6ebzdN5c7s8L9
+dX85XnCMHTa5h1zUZURVMwIL97xO/oSsDYF3fX3mU0QyDiA/FWmzvncI8Zu0iWAI
+oF9O8CGwWVbU4/iFHCWjbEYjQfqpw88l7qqstW5DkM7P6+T6GGJxQQQ6BmYhh7dn
+AS8U1xH6xGhV29ZI2sbbsNaWFlE6Z2Emaov8b1g7iQEEDmQ6xuV8KQ2Ykv9PQ/G1
+u6S9IqrmGqWJpkkub533uSa6/rVveTz6aq7G
+-----END CERTIFICATE-----
+    """
+)
+private fun parsePemCertificateChain(pemChain: List<String>): List<X509Certificate> {
+    require(pemChain.isNotEmpty()) { "No PEM certificates found" }
+    return pemChain.map { block ->
+        X509Certificate.decodeFromPem(block).getOrThrow()
+    }
+}